fix names, add icon, add description
