--- artifacttemplates/https%3A%2F%2Fust-quantil.github.io%2Fartifacttemplates/PlanQK-Platform-IA_w1-wip1/files/io_github_ust-quantil_artifactTemplates_PlanQK-Platform__ContainerInterface.war
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a1d855fc9e56648edb0b76856685f132ee169485170445ace802dbeb430ce602
-size 48240021
