--- artifacttemplates/https%3A%2F%2Fust-quantil.github.io%2Fartifacttemplates/PlanQK-Platform-IA_w1-wip1/source/ia/src/main/java/org/opentosca/artifacttemplates/dockercontainer/DockerContainerConstants.java
@@ -1,13 +1,0 @@
-package org.opentosca.artifacttemplates.dockercontainer;
-
-public abstract class DockerContainerConstants {
-
-    // Namespace under which the SOAP service operates
-    public static final String NAMESPACE_URI = "http://artifacttemplates.opentosca.org";
-
-    // Port type to use for the SOAP service
-    public static final String PORT_TYPE_NAME = "io_github_ust-quantil_artifactTemplates_PlanQK-Platform_ContainerInterfacePort";
-
-    // Name of the XML schema file to use as basis for the WSDL generation
-    public static final String XSD_NAME = "containerManagementInterface.xsd";
-}
