--- artifacttemplates/https%3A%2F%2Fust-quantil.github.io%2Fartifacttemplates/PlanQK-Platform-IA_w1-wip1/source/ia/src/main/java/org/opentosca/artifacttemplates/dockercontainer/WebServiceConfig.java
@@ -1,41 +1,0 @@
-package org.opentosca.artifacttemplates.dockercontainer;
-
-import org.springframework.boot.web.servlet.ServletRegistrationBean;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.ws.config.annotation.EnableWs;
-import org.springframework.ws.config.annotation.WsConfigurerAdapter;
-import org.springframework.ws.transport.http.MessageDispatcherServlet;
-import org.springframework.ws.wsdl.wsdl11.DefaultWsdl11Definition;
-import org.springframework.xml.xsd.SimpleXsdSchema;
-import org.springframework.xml.xsd.XsdSchema;
-
-@EnableWs
-@Configuration
-public class WebServiceConfig extends WsConfigurerAdapter {
-
-    @Bean
-    public ServletRegistrationBean<MessageDispatcherServlet> messageDispatcherServlet(ApplicationContext applicationContext) {
-        MessageDispatcherServlet servlet = new MessageDispatcherServlet();
-        servlet.setApplicationContext(applicationContext);
-        servlet.setTransformWsdlLocations(true);
-        return new ServletRegistrationBean<>(servlet, "/*");
-    }
-
-    @Bean(name = DockerContainerConstants.PORT_TYPE_NAME)
-    public DefaultWsdl11Definition defaultWsdl11Definition(XsdSchema dockerInterfaceSchema) {
-        DefaultWsdl11Definition wsdl11Definition = new DefaultWsdl11Definition();
-        wsdl11Definition.setPortTypeName(DockerContainerConstants.PORT_TYPE_NAME);
-        wsdl11Definition.setLocationUri("/");
-        wsdl11Definition.setTargetNamespace(DockerContainerConstants.NAMESPACE_URI);
-        wsdl11Definition.setSchema(dockerInterfaceSchema);
-        return wsdl11Definition;
-    }
-
-    @Bean
-    public XsdSchema containerManagementInterfaceSchema() {
-        return new SimpleXsdSchema(new ClassPathResource(DockerContainerConstants.XSD_NAME));
-    }
-}
