+++ artifacttemplates/https%3A%2F%2Fust-quantil.github.io%2Fartifacttemplates/PlanQK-Service-IA_w1-wip1/files/io_github_ust-quantil_artifactTemplates_PlanQK-Service__ContainerInterface.war
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d353b07754605d51a7011f46640ff606b0f9dde9ae7035008c049fd97a02e959
+size 48235449
