+++ artifacttemplates/https%3A%2F%2Fust-quantil.github.io%2Fartifacttemplates/PlanQK-Service-IA_w1-wip1/files/io_github_ust-quantil_artifactTemplates_PlanQK-Service__ContainerInterface.war.mimetype
