+++ artifacttemplates/https%3A%2F%2Fust-quantil.github.io%2Fartifacttemplates/PlanQK-Service-IA_w1-wip1/source/ia/src/main/java/org/opentosca/artifacttemplates/WsdlFilter.java
