+++ artifacttemplates/https%3A%2F%2Fust-quantil.github.io%2Fartifacttemplates/PlanQK-Service-IA_w1-wip1/source/ia/src/main/java/org/opentosca/artifacttemplates/planqkservice/feign/CustomDecoder.java
@@ -1,0 +1,39 @@
+package org.opentosca.artifacttemplates.planqkservice.feign;
+
+import java.io.IOException;
+import java.lang.reflect.Type;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import de.stoneone.planqk.api.invoker.ApiResponseDecoder;
+import feign.Response;
+import feign.Util;
+
+/**
+ * Custom decoder to monkey-patch the generated decoder class. This is required to support byte[] and String return types.
+ */
+public class CustomDecoder extends ApiResponseDecoder {
+
+    public CustomDecoder(ObjectMapper mapper) {
+        super(mapper);
+    }
+
+    @Override
+    public Object decode(Response response, Type type) throws IOException {
+        if (response.status() != 404 && response.status() != 204) {
+            if (response.body() == null) {
+                return null;
+            }
+            // decode byte arrays
+            if (byte[].class.equals(type)) {
+                return Util.toByteArray(response.body().asInputStream());
+            }
+            // decode strings
+            if (String.class.equals(type)) {
+                return Util.toString(response.body().asReader(Util.UTF_8));
+            }
+        } else {
+            return Util.emptyValueOf(type);
+        }
+        return super.decode(response, type);
+    }
+}
