+++ nodetypeimplementations/https%3A%2F%2Fust-quantil.github.io%2Fnodetypeimplementations/PlanQK-Service-Implementation_w1-wip1/NodeTypeImplementation.tosca
@@ -2,7 +2,7 @@
 <Definitions xmlns="http://docs.oasis-open.org/tosca/ns/2011/12" xmlns:winery="http://www.opentosca.org/winery/extensions/tosca/2013/02/12" xmlns:selfservice="http://www.eclipse.org/winery/model/selfservice" xmlns:testwineryopentoscaorg="http://test.winery.opentosca.org" targetNamespace="https://ust-quantil.github.io/nodetypeimplementations" id="nodeTypeImplementationsQuantil-PlanQK-Service-Implementation_w1-wip1">
     <NodeTypeImplementation xmlns:nodeTypesQuantil="https://ust-quantil.github.io/nodetypes" targetNamespace="https://ust-quantil.github.io/nodetypeimplementations" name="PlanQK-Service-Implementation_w1-wip1" abstract="no" final="no" nodeType="nodeTypesQuantil:PlanQK-Service_w1-wip1">
         <ImplementationArtifacts>
-            <ImplementationArtifact xmlns:otatyIgeneral="http://opentosca.org/artifacttypes" xmlns:artifactTemplatesQuantil="https://ust-quantil.github.io/artifacttemplates" name="PlanQK-PlatformIA" interfaceName="http://opentosca.org/interfaces/pattern/container" artifactType="otatyIgeneral:WAR-Java17" artifactRef="artifactTemplatesQuantil:PlanQK-Platform-IA_w1-wip1"/>
+            <ImplementationArtifact xmlns:otatyIgeneral="http://opentosca.org/artifacttypes" xmlns:artifactTemplatesQuantil="https://ust-quantil.github.io/artifacttemplates" name="PlanQK-Service-IA" interfaceName="http://opentosca.org/interfaces/pattern/container" artifactType="otatyIgeneral:WAR-Java17" artifactRef="artifactTemplatesQuantil:PlanQK-Service-IA_w1-wip1"/>
         </ImplementationArtifacts>
         <DeploymentArtifacts>
             <DeploymentArtifact xmlns:ToscaBaseTypes="http://docs.oasis-open.org/tosca/ns/2011/12/ToscaBaseTypes" xmlns:artifactTemplatesQuantil="https://ust-quantil.github.io/artifacttemplates" name="UserCode" artifactType="ToscaBaseTypes:ArchiveArtifact" artifactRef="artifactTemplatesQuantil:PlanQK-Service-User-Code_-DA-w1-wip1"/>
