+++ servicetemplates/http%3A%2F%2Fopentosca.org%2Fservicetemplates/PlanQK-Service-Example_w1-wip1/SELFSERVICE-Metadata/data.xml.mimetype
