+++ servicetemplates/http%3A%2F%2Fopentosca.org%2Fservicetemplates/PlanQK-Service-Example_w1-wip1/SELFSERVICE-Metadata/icon.jpg.mimetype
@@ -1,0 +1,1 @@
+image/png
