+++ servicetemplates/http%3A%2F%2Fopentosca.org%2Fservicetemplates/PlanQK-Service-Example_w1-wip1/ServiceTemplate.tosca
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<Definitions xmlns="http://docs.oasis-open.org/tosca/ns/2011/12" xmlns:winery="http://www.opentosca.org/winery/extensions/tosca/2013/02/12" xmlns:selfservice="http://www.eclipse.org/winery/model/selfservice" xmlns:testwineryopentoscaorg="http://test.winery.opentosca.org" targetNamespace="http://opentosca.org/servicetemplates" id="otsteIgeneral-PlanQK-Service-Example_w1-wip1">
+    <ServiceTemplate name="PlanQK-Service-Example_w1-wip1" targetNamespace="http://opentosca.org/servicetemplates" id="PlanQK-Service-Example_w1-wip1">
+        <TopologyTemplate>
+            <NodeTemplate xmlns:nodeTypesQuantil="https://ust-quantil.github.io/nodetypes" name="PlanQK-Platform_w1-wip1" minInstances="1" maxInstances="1" type="nodeTypesQuantil:PlanQK-Platform_w1-wip1" id="PlanQK-Platform_w1-wip1_0" winery:x="352" winery:y="749">
+                <Properties>
+                    <Properties xmlns="http://www.example.org">
+                        <PlanqkApiKey>get_input: PlanqkApiKey</PlanqkApiKey>
+                    </Properties>
+                </Properties>
+            </NodeTemplate>
+            <NodeTemplate xmlns:nodeTypesQuantil="https://ust-quantil.github.io/nodetypes" name="PlanQK-Service_w1-wip1" minInstances="1" maxInstances="1" type="nodeTypesQuantil:PlanQK-Service_w1-wip1" id="PlanQK-Service_w1-wip1_0" winery:x="352" winery:y="375">
+                <Properties>
+                    <Properties xmlns="http://www.example.org">
+                        <ServiceName>TestService</ServiceName>
+                        <ServiceDescription>test service</ServiceDescription>
+                        <QuantumBackend>NONE</QuantumBackend>
+                    </Properties>
+                </Properties>
+                <DeploymentArtifacts>
+                    <DeploymentArtifact xmlns:ToscaBaseTypes="http://docs.oasis-open.org/tosca/ns/2011/12/ToscaBaseTypes" xmlns:artifactTemplatesQuantil="https://ust-quantil.github.io/artifacttemplates" name="UserCode" artifactType="ToscaBaseTypes:ArchiveArtifact" artifactRef="artifactTemplatesQuantil:PlanQK-Service-User-Code_-DA-w1-wip1"/>
+                </DeploymentArtifacts>
+            </NodeTemplate>
+            <NodeTemplate xmlns:nodeTypesQuantil="https://ust-quantil.github.io/nodetypes" name="PlanQK-ServiceInstanceInfo_w1-wip1" minInstances="1" maxInstances="1" type="nodeTypesQuantil:PlanQK-ServiceInstanceInfo_w1-wip1" id="PlanQK-ServiceInstanceInfo_w1-wip1_0" winery:x="352" winery:y="112">
+                <Properties>
+                    <PlanQK-ServiceInfo_w1-wip1 xmlns="https://ust-quantil.github.io/nodetypes">
+                        <ServiceID/>
+                    </PlanQK-ServiceInfo_w1-wip1>
+                </Properties>
+            </NodeTemplate>
+            <RelationshipTemplate xmlns:ToscaBaseTypes="http://docs.oasis-open.org/tosca/ns/2011/12/ToscaBaseTypes" name="HostedOn" type="ToscaBaseTypes:HostedOn" id="con_HostedOn_0">
+                <SourceElement ref="PlanQK-Service_w1-wip1_0"/>
+                <TargetElement ref="PlanQK-Platform_w1-wip1_0"/>
+            </RelationshipTemplate>
+            <RelationshipTemplate xmlns:ToscaBaseTypes="http://docs.oasis-open.org/tosca/ns/2011/12/ToscaBaseTypes" name="HostedOn" type="ToscaBaseTypes:HostedOn" id="con_HostedOn_1">
+                <SourceElement ref="PlanQK-ServiceInstanceInfo_w1-wip1_0"/>
+                <TargetElement ref="PlanQK-Service_w1-wip1_0"/>
+            </RelationshipTemplate>
+        </TopologyTemplate>
+    </ServiceTemplate>
+</Definitions>
