--- servicetemplates/http%3A%2F%2Fopentosca.org%2Fservicetemplates/PlanQKTest_w1-wip1/SELFSERVICE-Metadata/data.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<selfservice:Application xmlns="http://docs.oasis-open.org/tosca/ns/2011/12" xmlns:winery="http://www.opentosca.org/winery/extensions/tosca/2013/02/12" xmlns:yml="http://docs.oasis-open.org/tosca/ns/simple/yaml/1.3" xmlns:selfservice="http://www.eclipse.org/winery/model/selfservice" xmlns:researchobject="http://www.eclipse.org/winery/model/researchobject" xmlns:testwineryopentoscaorg="http://test.winery.opentosca.org">
-    <selfservice:version>1.0</selfservice:version>
-    <selfservice:authors>Winery</selfservice:authors>
-</selfservice:Application>
